ueu/libs/angular/shared/core/notifications/src/lib/components/confirmation-modal/confirmation-modal.component.ts
import { Component, Inject } from '@angular/core';
import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';

@Component({
    selector: 'picpay-confirmation-modal',
    templateUrl: './confirmation-modal.component.html',
    styleUrls: ['./confirmation-modal.component.scss'],
})
export class ConfirmationModalComponent {
    constructor(
        private dialogRef: MatDialogRef<ConfirmationModalComponent>,
        @Inject(MAT_DIALOG_DATA) public data: { title: string; message: string },
    ) {}

    onClose(confirm: boolean): void {
        this.dialogRef.close({ confirm });
    }
}
